rate-limited and secured endpoints

+++ db/questionOperations.js
@@ -12,10 +12,25 @@ export async function getQuestionsFromSetId(id) {
       },
     }
   )
-  return user
+  return user;
 } 
 
-export async function deleteQuestion (id) {
+export async function getQuestion (id) {
+  const getQuestion = await prisma.question.findFirstOrThrow(
+    {
+      where: {
+        id
+      }
+    }
+  )
+  return getQuestion;
+}
+
+export async function deleteQuestion (id, userId) {
+  const findQuestion = await getQuestion(id);
+  if (findQuestion.userId !== userId) {
+    throw new Error('User is not authorized...')
+  }
   const deleteQuestion = await prisma.question.delete({
     where: {
       id
@@ -24,10 +39,11 @@ export async function deleteQuestion (id) {
   return deleteQuestion;
 }
 
-export async function insertQuestion (question) {
+export async function insertQuestion (question, userId) {
   const questionCreate = await prisma.question.create({
     data: {
       question: question.question,
+      userId,
       set: {
         connect: { id: question.setId }
       }
@@ -36,7 +52,11 @@ export async function insertQuestion (question) {
   return questionCreate;
 }
 
-export async function updateQuestion (questionInfo) {
+export async function updateQuestion (questionInfo, userId) {
+  const findQuestion = await getQuestion(questionInfo.id);
+  if (findQuestion.userId !== userId) {
+    throw new Error('User is not authorized...')
+  }
   const updateQuestion = await prisma.question.update({
     where: {
       id: questionInfo.id
